you can update your reservation, now

+++ frontend/src/app/views/reservations-page/reservations-page.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { Router } from '@angular/router';
 import { Reservations } from 'src/app/shared/models/reservations';
 import { ReservationsService } from 'src/app/shared/services/reservations.service';
 import { StorageService } from 'src/app/shared/services/storage.service';
@@ -13,7 +14,7 @@ export class ReservationsPageComponent {
 
   reservations: Array<Reservations> = []
 
-  constructor(private reservationsService: ReservationsService, private storage: StorageService) { }
+  constructor(private reservationsService: ReservationsService, private storage: StorageService, private router: Router) { }
 
   ngOnInit() {
     let data = this.reservationsService.getAllReservations();
@@ -28,26 +29,31 @@ export class ReservationsPageComponent {
     })
   }
 
-  checkOrder(id: number){
-    Swal.fire({
-      title: 'Do you want to delete your reservation?',
-      text: "You won't be able to revert this!",
-      icon: 'warning',
-      showCancelButton: true,
-      confirmButtonColor: '#DAD2BC',
-      cancelButtonColor: '#69747C',
-      confirmButtonText: 'Yes'
-    }).then((result) => {
-      if (result.isConfirmed) {
-        this.reservationsService.deleteReservation(id);
-        Swal.fire(
-          'Done!',
-          'Your reservation has been deleted correctly.',
-          'success'
-        ).then(function () {
-          window.location.reload()
-        })
-      }
-    })
+  checkOrder(id: number, index: number) {
+    console.log(this.reservations[index].startDate);
+    if (new Date(this.reservations[index].startDate).toISOString() == new Date().toISOString() || new Date(this.reservations[index].startDate).toISOString() < new Date().toISOString()) {
+      Swal.fire({
+        title: 'Do you want to delete your reservation?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#DAD2BC',
+        cancelButtonColor: '#69747C',
+        confirmButtonText: 'Yes'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          this.reservationsService.deleteReservation(id);
+          Swal.fire(
+            'Done!',
+            'Your reservation has been deleted correctly.',
+            'success'
+          ).then(function () {
+            window.location.reload()
+          })
+        }
+      })
+    } else {
+    this.router.navigateByUrl(`/edit-order/${id}`);
+    }
   }
 }
